+++ influxql/select.go
@@ -748,7 +748,7 @@ func (b *exprIteratorBuilder) buildCallIterator(expr *Call) (Iterator, error) {
 		opt.Interval = Interval{}
 
 		return newHoltWintersIterator(input, opt, int(h.Val), int(m.Val), includeFitData, interval)
-	case "derivative", "non_negative_derivative", "difference", "moving_average", "elapsed":
+	case "derivative", "non_negative_derivative", "difference", "non_negative_difference", "moving_average", "elapsed":
 		opt := b.opt
 		if !opt.Interval.IsZero() {
 			if opt.Ascending {
@@ -772,8 +772,9 @@ func (b *exprIteratorBuilder) buildCallIterator(expr *Call) (Iterator, error) {
 		case "elapsed":
 			interval := opt.ElapsedInterval()
 			return newElapsedIterator(input, opt, interval)
-		case "difference":
-			return newDifferenceIterator(input, opt)
+		case "difference", "non_negative_difference":
+			isNonNegative := (expr.Name == "non_negative_difference")
+			return newDifferenceIterator(input, opt, isNonNegative)
 		case "moving_average":
 			n := expr.Args[1].(*IntegerLiteral)
 			if n.Val > 1 && !b.opt.Interval.IsZero() {
@@ -794,6 +795,15 @@ func (b *exprIteratorBuilder) buildCallIterator(expr *Call) (Iterator, error) {
 			return nil, err
 		}
 		return newCumulativeSumIterator(input, opt)
+	case "integral":
+		opt := b.opt
+		opt.Ordered = true
+		input, err := buildExprIterator(expr.Args[0].(*VarRef), b.ic, b.sources, opt, false)
+		if err != nil {
+			return nil, err
+		}
+		interval := opt.IntegralInterval()
+		return newIntegralIterator(input, opt, interval)
 	}
 
 	itr, err := func() (Iterator, error) {
@@ -1183,6 +1193,43 @@ func buildRHSTransformIterator(lhs Iterator, rhs Literal, op Token, opt Iterator
 				return bp
 			},
 		}, nil
+	case func(bool, bool) bool:
+		var input BooleanIterator
+		switch lhs := lhs.(type) {
+		case BooleanIterator:
+			input = lhs
+		default:
+			return nil, fmt.Errorf("type mismatch on LHS, unable to use %T as an BooleanIterator", lhs)
+		}
+
+		var val bool
+		switch rhs := rhs.(type) {
+		case *BooleanLiteral:
+			val = rhs.Val
+		default:
+			return nil, fmt.Errorf("type mismatch on RHS, unable to use %T as an BooleanLiteral", rhs)
+		}
+		return &booleanTransformIterator{
+			input: input,
+			fn: func(p *BooleanPoint) *BooleanPoint {
+				if p == nil {
+					return nil
+				}
+
+				bp := &BooleanPoint{
+					Name: p.Name,
+					Tags: p.Tags,
+					Time: p.Time,
+					Aux:  p.Aux,
+				}
+				if p.Nil {
+					bp.Nil = true
+				} else {
+					bp.Value = fn(p.Value, val)
+				}
+				return bp
+			},
+		}, nil
 	}
 	return nil, fmt.Errorf("unable to construct rhs transform iterator from %T and %T", lhs, rhs)
 }
@@ -1362,6 +1409,43 @@ func buildLHSTransformIterator(lhs Literal, rhs Iterator, op Token, opt Iterator
 				return bp
 			},
 		}, nil
+	case func(bool, bool) bool:
+		var input BooleanIterator
+		switch rhs := rhs.(type) {
+		case BooleanIterator:
+			input = rhs
+		default:
+			return nil, fmt.Errorf("type mismatch on RHS, unable to use %T as an BooleanIterator", rhs)
+		}
+
+		var val bool
+		switch lhs := lhs.(type) {
+		case *BooleanLiteral:
+			val = lhs.Val
+		default:
+			return nil, fmt.Errorf("type mismatch on LHS, unable to use %T as a BooleanLiteral", lhs)
+		}
+		return &booleanTransformIterator{
+			input: input,
+			fn: func(p *BooleanPoint) *BooleanPoint {
+				if p == nil {
+					return nil
+				}
+
+				bp := &BooleanPoint{
+					Name: p.Name,
+					Tags: p.Tags,
+					Time: p.Time,
+					Aux:  p.Aux,
+				}
+				if p.Nil {
+					bp.Nil = true
+				} else {
+					bp.Value = fn(val, p.Value)
+				}
+				return bp
+			},
+		}, nil
 	}
 	return nil, fmt.Errorf("unable to construct lhs transform iterator from %T and %T", lhs, rhs)
 }
@@ -1438,9 +1522,19 @@ func buildTransformIterator(lhs Iterator, rhs Iterator, op Token, opt IteratorOp
 		}
 		right, ok := rhs.(IntegerIterator)
 		if !ok {
-			return nil, fmt.Errorf("type mismatch on LHS, unable to use %T as a IntegerIterator", rhs)
+			return nil, fmt.Errorf("type mismatch on RHS, unable to use %T as a IntegerIterator", rhs)
 		}
 		return newIntegerBooleanExprIterator(left, right, opt, fn), nil
+	case func(bool, bool) bool:
+		left, ok := lhs.(BooleanIterator)
+		if !ok {
+			return nil, fmt.Errorf("type mismatch on LHS, unable to use %T as a BooleanIterator", lhs)
+		}
+		right, ok := rhs.(BooleanIterator)
+		if !ok {
+			return nil, fmt.Errorf("type mismatch on RHS, unable to use %T as a BooleanIterator", rhs)
+		}
+		return newBooleanExprIterator(left, right, opt, fn), nil
 	}
 	return nil, fmt.Errorf("unable to construct transform iterator from %T and %T", lhs, rhs)
 }
@@ -1487,6 +1581,8 @@ func binaryExprFunc(typ1 DataType, typ2 DataType, op Token) interface{} {
 		default:
 			fn = integerBinaryExprFunc(op)
 		}
+	case Boolean:
+		fn = booleanBinaryExprFunc(op)
 	}
 	return fn
 }
@@ -1546,6 +1642,12 @@ func integerBinaryExprFunc(op Token) interface{} {
 			}
 			return lhs % rhs
 		}
+	case BITWISE_AND:
+		return func(lhs, rhs int64) int64 { return lhs & rhs }
+	case BITWISE_OR:
+		return func(lhs, rhs int64) int64 { return lhs | rhs }
+	case BITWISE_XOR:
+		return func(lhs, rhs int64) int64 { return lhs ^ rhs }
 	case EQ:
 		return func(lhs, rhs int64) bool { return lhs == rhs }
 	case NEQ:
@@ -1558,6 +1660,18 @@ func integerBinaryExprFunc(op Token) interface{} {
 		return func(lhs, rhs int64) bool { return lhs > rhs }
 	case GTE:
 		return func(lhs, rhs int64) bool { return lhs >= rhs }
+	}
+	return nil
+}
+
+func booleanBinaryExprFunc(op Token) interface{} {
+	switch op {
+	case BITWISE_AND:
+		return func(lhs, rhs bool) bool { return lhs && rhs }
+	case BITWISE_OR:
+		return func(lhs, rhs bool) bool { return lhs || rhs }
+	case BITWISE_XOR:
+		return func(lhs, rhs bool) bool { return lhs != rhs }
 	}
 	return nil
 }